refactor: change name of components and variable
