+++ app/chat/components/utilityWindow/createGroupWindow/createGroupWindow.tsx
@@ -1,0 +1,7 @@
+export function CreateGroupWindow() {
+  return (
+    <div>
+      <p>This is create group</p>
+    </div>
+  );
+}
