--- app/chat/components/utilityWindow/searchWindow/searchWindow.tsx
@@ -1,7 +1,0 @@
-export function SearchWindow() {
-  return (
-    <div>
-      <p>This is search</p>
-    </div>
-  );
-}
